+++ frontend/src/components/SideBarFooter.jsx
@@ -6,10 +6,17 @@ import { useNavigate } from 'react-router-dom';
 import { useLogoutMutation } from '../slices/usersApiSlice';
 import { logout } from '../slices/authSlice';
 import { useEffect, useState } from "react";
+import UserRole from "../utils/userRole";
 
 const SideBarFooter = () => {
     const [fullName, setFullName] = useState('');
     const { userInfo } = useSelector((state) => state.auth);
+
+    const data = UserRole();
+
+    const isAdmin = data?.role === 'admin';
+    const isEmployee = data?.role === 'empleado';
+    const isClient = data?.role === 'cliente';
 
     const dispatch = useDispatch();
     const navigate = useNavigate();
@@ -38,7 +45,7 @@ const SideBarFooter = () => {
                     <Sidebar.Nav.Title id='username'>{fullName}</Sidebar.Nav.Title>
                 </Dropdown.Toggle>
                 <Dropdown.Menu>
-                    <Dropdown.Item href="/profile">Perfil</Dropdown.Item>
+                    <Dropdown.Item href={isClient ? '/client/profile' : '/bank/profile'}>Perfil</Dropdown.Item>
                     <Dropdown.Item onClick={logoutHandler}>Cerrar sesion</Dropdown.Item>
                 </Dropdown.Menu>
             </Dropdown>
