+++ frontend/src/views/bank/Dashboard.jsx
@@ -1,17 +1,17 @@
 import React, { useEffect } from 'react';
-import CardContainer from '../components/CardContainer';
-import { useShowAccountsQuery } from '../slices/accountApiSlice';
-import { useShowAllOperationsQuery } from '../slices/operationApiSlice';
+import CardContainer from '../../components/CardContainer';
+import { useShowAccountsQuery } from '../../slices/accountApiSlice';
+import { useShowAllOperationsQuery } from '../../slices/operationApiSlice';
 import { LinkContainer } from 'react-router-bootstrap';
 import { Link, useNavigate } from 'react-router-dom';
 import { Button, Nav } from 'react-bootstrap';
-import Loader from '../components/Loader';
+import Loader from '../../components/Loader';
 import { toast } from 'react-toastify';
-import ContentBox from '../components/ContentBox';
+import ContentBox from '../../components/ContentBox';
 import { FaChevronCircleRight } from "react-icons/fa";
-import ImageContainer from '../components/ImageContainer';
-import useCheckCookies from '../utils/useCheckCookies';
-import useSessionTimeout from '../utils/useSessionTimeout';
+import ImageContainer from '../../components/ImageContainer';
+import useCheckCookies from '../../utils/useCheckCookies';
+import useSessionTimeout from '../../utils/useSessionTimeout';
 
 const Home = () => {
   useCheckCookies();
