+++ frontend/src/views/client/AccountSummary.jsx
@@ -1,15 +1,15 @@
 import React, { useEffect, useState } from 'react';
-import CardContainer from '../components/CardContainer';
-import useCheckCookies from '../utils/useCheckCookies';
-import BoxContainer from '../components/BoxContainer';
-import useSessionTimeout from '../utils/useSessionTimeout';
+import CardContainer from '../../components/CardContainer';
+import useCheckCookies from '../../utils/useCheckCookies';
+import BoxContainer from '../../components/BoxContainer';
+import useSessionTimeout from '../../utils/useSessionTimeout';
 import { Button, Col, Form } from 'react-bootstrap';
 import Select from 'react-select';
-import { useShowAccountsQuery } from '../slices/accountApiSlice';
-import UserAccounts from '../utils/userAccounts';
-import AccountSummaryPDF from '../components/AccountSummaryPDF';
+import { useShowAccountsQuery } from '../../slices/accountApiSlice';
+import UserAccounts from '../../utils/userAccounts';
+import AccountSummaryPDF from '../../components/AccountSummaryPDF';
 import { pdf } from '@react-pdf/renderer';
-import { useShowAccountOperationsQuery } from '../slices/operationApiSlice';
+import { useShowAccountOperationsQuery } from '../../slices/operationApiSlice';
 import { toast } from 'react-toastify';
 import { useSelector } from 'react-redux';
