+++ frontend/src/views/client/Profile.jsx
@@ -1,11 +1,11 @@
 import { useState, useEffect } from 'react';
 import { Form, Button } from 'react-bootstrap';
 import { useDispatch, useSelector } from 'react-redux';
-import FormContainer from '../components/FormContainer';
+import FormContainer from '../../components/FormContainer';
 import { toast } from 'react-toastify';
-import Loader from '../components/Loader';
-import { useUpdateUserMutation } from '../slices/usersApiSlice';
-import { setCredentials } from '../slices/authSlice';
+import Loader from '../../components/Loader';
+import { useUpdateUserMutation } from '../../slices/usersApiSlice';
+import { setCredentials } from '../../slices/authSlice';
 
 const Profile = () => {
     const [email, setEmail] = useState('');
